pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
[project]
name = "finlab"
version = "0.1.0"
description = "Extractor y simulador financiero multi-API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.9",
  "pandas>=2.0",
  "numpy>=1.24",
  "python-dotenv>=1.0",
  "requests>=2.31",
  "matplotlib>=3.8",
  "pyarrow>=14.0"   
]
[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
finlab = "finlab.cli:app"
